Added example file; ironed out bugs in ENVICamera

+++ src/frameview_widget.cpp
@@ -203,9 +203,10 @@ void frameview_widget::handleNewFrame()
     if (!this->isHidden() && frame_handler->running()) {
 
         std::vector<float>image_data{(frame_handler->*p_getFrame)()};
-        for (unsigned int col = 0; col < frWidth; col++) {
-            for (unsigned int row = 0; row < frHeight; row++ ) {
-                colorMap->data()->setCell(col, row, image_data[row * frWidth + col]); // y-axis NOT reversed
+        for (int col = 0; col < frWidth; col++) {
+            for (int row = 0; row < frHeight; row++ ) {
+                colorMap->data()->setCell(col, row,
+                                          double(image_data[size_t(row * frWidth + col)])); // y-axis NOT reversed
             }
         }
         qcp->replot();
